rename contract_types.hpp to operators.hpp
