+++ include/contracts_lite/simple_violation_handler.hpp
@@ -25,7 +25,7 @@
 #include <stdexcept>
 #include <string>
 
-#include "contracts_lite/contract_types.hpp"
+#include "contracts_lite/operators.hpp"
 
 namespace contracts_lite {
